// src/api/model/eventCreationDTO.ts
/**
 * RoomReservation
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface EventCreationDTO { 
    title?: string;
    description?: string;
    start?: string;
    end?: string;
    isPublic?: boolean;
}